Add C++ interfaces to Fortran code.
The only business code implemented on the C++ side is the simulation loop handling different configurations of inputs. For this purpose, a Mask class is implemented to keep track of which points are exlcuded during a given run. This is necessary because the linear algebra objects from Armadillo don't support boolean values.