Fix some issues with the cli progress bar
Replace calls to REprintf (which prints to R's stderr) by Rprintf. The former was simply copied from the template in some RcppProgress doc but doesn't seem necessary, and breaks testing because the output cannot be captured.

Remove the call to sprintf. R CMD check complains about this even though it was outputting to a variable.