Remove package config class and transfer its functionality to the RNG class. Rename it in anticipation of package renaming. Fix the RNG mechanism based on the TRNG library to function correctly.
The rationale for having a separate configuration singleton was based on the anticipation that it would be dealing with other options irrelevant to the RNG, but this has not materialised. Having to maintain and communicate between between two singletons was becoming burdensome and error-prone.

Our understanding of how TRNG works was erroneous. It seems we need to give every thread its own copy of the engine and advance those copies by a thread-dependent step to obtain independent streams.